Signout Api

+++ admin-app/src/actions/auth.action.js
@@ -51,9 +51,20 @@ export const isUserLoggedIn = () => {
 
 export const signout = () => {
     return async dispatch => {
-        localStorage.clear();
         dispatch({
             type: authConstants.LOGOUT_REQUEST
         })
+        const res = await axiosInstance.post(`/admin/signout`);
+        if(res.status === 200) {
+            localStorage.clear();
+            dispatch({
+                type: authConstants.LOGOUT_SUCCESS
+            });
+        } else {
+            dispatch({
+                type: authConstants.LOGOUT_FAILURE,
+                payload: { error: res.data.error }
+            })
+        }
     }
 }
